+++ common/cpp/src/logger/logger.cpp
@@ -38,7 +38,7 @@ LogLevel StringToLogLevel(const std::string& name, Error* err) {
 
 template<typename ... Args>
 std::string string_format( const std::string& format, Args ... args ) {
-    size_t size = snprintf( nullptr, 0, format.c_str(), args ... ) + 1;
+    size_t size = static_cast<size_t>(snprintf( nullptr, 0, format.c_str(), args ... ) + 1);
     std::unique_ptr<char[]> buf( new char[ size ] );
     snprintf( buf.get(), size, format.c_str(), args ... );
     return std::string( buf.get(), buf.get() + size - 1 );
